Project/Tollplaza/requirements.txt
autopy (0.51)
cycler (0.10.0)
Django (1.8)
django-filter (1.0.2)
djangorestframework (3.6.2)
EasyProcess (0.2.3)
functools32 (3.2.3.post2)
Markdown (2.6.8)
mysql-python (1.2.5)
numpy (1.12.0)
olefile (0.44)
opencv-python (3.2.0.6)
passlib (1.7.1)
Pillow (4.0.0)
pip (9.0.1)
pyBarcode (0.7)
pygame (1.9.3)
PyMySQL (0.7.9)
pyping (0.0.5)
pypiwin32 (219)
pyscreenshot (0.4.2)
python-dateutil (2.6.0)
python-mimeparse (1.6.0)
pytz (2016.10)
setuptools (32.0.0)
six (1.10.0)
virtualenv (15.1.0)
wheel (0.29.0)
